move load_config to config.py
